Start refactoring the TUI
