+++ libtiny_tui/Cargo.toml
@@ -18,6 +18,7 @@ name = "libtiny_tui"
 path = "src/lib.rs"
 
 [dependencies]
+futures-util-preview = "0.3.0-alpha.18"
 libtiny_ui = { path = "../libtiny_ui" }
 notify-rust = "3"
 serde = { version = "1.0.8", features = ["derive"] }
@@ -25,6 +26,7 @@ tempfile = "3.0.3"
 term_input = { path = "../term_input" }
 termbox_simple = { path = "../termbox", version = "0.2.3" }
 time = "0.1"
+tokio = { git = "https://github.com/tokio-rs/tokio.git", features = ["timer"] }
 
 [dev-dependencies]
 mio = "0.6.9"
